[3.11] gh-108303: Fix and move `badsyntax_pep3120.py` (GH-109513)
(cherry picked from commit 4dd47c63a97b3c39cd964ad12431fcdaf76dc823)
